Add files via upload
These SQL scripts showcase the initial data cleaning and creation of the tables that would subsequently be imported into PowerBi to create an interactive dashboard as requested. The project aimed to gain insights into the sales trends of a particular company by analyzing its data on sales, customers, products, and transaction dates.